+++ utils/montecarlo.go
@@ -5,15 +5,15 @@ import (
 )
 
 type MCF struct {
-	Function    func(float64) float64 //被积函数
-	AreaLeft    float64               //左边界
-	AreaRight   float64               //右边界
-	Integration float64               //积分结果
+	Function    func(float32) float32 //被积函数
+	AreaLeft    float32               //左边界
+	AreaRight   float32               //右边界
+	Integration float32               //积分结果
 	Time        int                   //随机次数，次数越大精度越高，性能开销越大
 }
 
 // 利用大数统计的思想近似求解函数最大值
-func (mc *MCF) getMax() (maxNum float64) {
+func (mc *MCF) getMax() (maxNum float32) {
 	var time = 10000 //默认次数10000次
 	maxNum = mc.Function(mc.AreaLeft)
 	d := mc.AreaRight - mc.AreaLeft
@@ -21,7 +21,7 @@ func (mc *MCF) getMax() (maxNum float64) {
 		time = mc.Time
 	}
 	for i := 0; i < time; i++ {
-		x := mc.AreaLeft + d*rand.Float64()
+		x := mc.AreaLeft + d*rand.Float32()
 		if mc.Function(x) > maxNum {
 			maxNum = mc.Function(x)
 		}
@@ -30,7 +30,7 @@ func (mc *MCF) getMax() (maxNum float64) {
 }
 
 // 利用大数统计的思想近似求解函数最小值
-func (mc *MCF) getMin() (minNum float64) {
+func (mc *MCF) getMin() (minNum float32) {
 	var time = 10000 //默认次数10000次
 	minNum = mc.Function(mc.AreaLeft)
 	d := mc.AreaRight - mc.AreaLeft
@@ -38,7 +38,7 @@ func (mc *MCF) getMin() (minNum float64) {
 		time = mc.Time
 	}
 	for i := 0; i < time; i++ {
-		x := mc.AreaLeft + d*rand.Float64()
+		x := mc.AreaLeft + d*rand.Float32()
 		if mc.Function(x) < minNum {
 			minNum = mc.Function(x)
 		}
@@ -47,17 +47,17 @@ func (mc *MCF) getMin() (minNum float64) {
 }
 
 // DeIntCalc 利用蒙特卡罗方法，通过面积比例关系近似求解函数定积分
-func (mc *MCF) DeIntCalc() float64 {
+func (mc *MCF) DeIntCalc() float32 {
 	var time = 10000
-	var num float64 = 0
+	var num float32 = 0
 	d1 := mc.AreaRight - mc.AreaLeft
 	d2 := mc.getMax() - mc.getMin()
 	if mc.Time > 0 {
 		time = mc.Time
 	}
 	for i := 0; i < time; i++ {
-		x := mc.AreaLeft + d1*rand.Float64()
-		y := mc.getMin() + d2*rand.Float64()
+		x := mc.AreaLeft + d1*rand.Float32()
+		y := mc.getMin() + d2*rand.Float32()
 		if mc.Function(x) > 0 {
 			if y < mc.Function(x) && y > 0 {
 				num++
@@ -69,7 +69,7 @@ func (mc *MCF) DeIntCalc() float64 {
 		}
 	}
 	s := d1 * d2
-	result := s * (num / float64(time))
+	result := s * (num / float32(time))
 	mc.Integration = result
 	return result
 }
